Work around Check Point VPN gateway IP mismatch
Allow connection to a Check Point IKEv1 gateway which IP is
misconfigured and causes the Subject Alternative Name (SAN) IP Address
in the certificate extension to not match the actual IP of the gateway.

By specifying in strongSwan configuration of the remote the id to be
equal to either the actual Subject or the (incorrect) SAN IP Address of
the certificate received from the gateway, the workaround makes
strongSwan use that certificate instead of deleting the connection and
complaining it cannot find a trusted RSA public key for the gateway.